backend/requirements.txt
twilio==7.16.0
flask==2.3.2
flask-cors==4.0.0
python-dotenv==1.0.0
requests==2.31.0